+++ packages/extension/src/components/molecules/TokenDisplay/RenderTokenIcon/index.ts
@@ -1,0 +1,1 @@
+export * from './RenderTokenIcon';
